--- Scenchive/src/main/java/com/example/scenchive/comment/controller/CommentController.java
@@ -1,5 +1,0 @@
-package com.example.scenchive.comment.controller;
-
-
-public class CommentController {
-}
