--- Scenchive/src/main/java/com/example/scenchive/comment/dto/CommentSaveDto.java
@@ -1,10 +1,0 @@
-package com.example.scenchive.comment.dto;
-
-import com.example.scenchive.comment.Comment;
-
-public class CommentSaveDto (String content){
-
-    public Comment toEntity() {
-        return Comment.builder().content(content).build();
-    }
-}
