+++ Scenchive/src/main/java/com/example/scenchive/domain/board/repository/boardType.java
@@ -1,0 +1,30 @@
+package com.example.scenchive.domain.board.repository;
+
+import jakarta.persistence.*;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Getter
+@NoArgsConstructor
+@Entity
+@Table(name="boardType")
+public class boardType {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String boardtype_name;
+
+    @OneToMany(mappedBy = "boardtype")
+    private List<Board> boards=new ArrayList<>();
+    //다 쪽에서 FK를 가지고 일 쪽에서 mappedby, list객체 가짐
+
+    @Builder
+    public boardType(String boardtype_name) {
+        this.boardtype_name = boardtype_name;
+    }
+}
