+++ Scenchive/src/main/java/com/example/scenchive/domain/comment/controller/CommentController.java
@@ -1,0 +1,5 @@
+package com.example.scenchive.domain.comment.controller;
+
+
+public class CommentController {
+}
