+++ Scenchive/src/main/java/com/example/scenchive/domain/comment/dto/CommentSaveDto.java
@@ -1,0 +1,10 @@
+package com.example.scenchive.domain.comment.dto;
+
+import com.example.scenchive.domain.comment.repository.Comment;
+
+public class CommentSaveDto (String content){
+
+    public Comment toEntity() {
+        return Comment.builder().content(content).build();
+    }
+}
