+++ Scenchive/src/main/java/com/example/scenchive/domain/comment/exception/CommentException.java
@@ -1,0 +1,18 @@
+package com.example.scenchive.domain.comment.exception;
+
+import com.example.scenchive.global.exception.BaseException;
+import com.example.scenchive.global.exception.BaseExceptionType;
+
+public class CommentException extends BaseException {
+
+    private BaseExceptionType baseExceptionType;
+
+    public CommentException(BaseExceptionType baseExceptionType) {
+        this.baseExceptionType = baseExceptionType;
+    }
+
+    @Override
+    public BaseExceptionType getExceptionType() {
+        return this.baseExceptionType;
+    }
+}
