+++ Scenchive/src/main/java/com/example/scenchive/domain/comment/repository/CommentRepository.java
@@ -1,0 +1,6 @@
+package com.example.scenchive.domain.comment.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+}
