+++ Scenchive/src/main/java/com/example/scenchive/domain/member/repository/MemberRepository.java
@@ -1,0 +1,12 @@
+package com.example.scenchive.domain.member.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    List<Member> findByName(String name);
+
+    Optional<Member> findByEmail(String email);
+}
