+++ Scenchive/src/main/java/com/example/scenchive/domain/member/service/NotCorrespondingEmailException.java
@@ -1,0 +1,10 @@
+package com.example.scenchive.domain.member.service;
+
+public class NotCorrespondingEmailException extends Exception {
+    public NotCorrespondingEmailException() {
+    }
+
+    public NotCorrespondingEmailException(String message) {
+        super(message);
+    }
+}
