--- Scenchive/src/main/java/com/example/scenchive/member/dto/LoginForm.java
@@ -1,16 +1,0 @@
-package com.example.scenchive.member.dto;
-
-import jakarta.validation.constraints.Email;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class LoginForm {
-    @Email
-    private String email;
-
-    private int password;
-}
