--- Scenchive/src/main/java/com/example/scenchive/member/exception/MemberExceptionType.java
@@ -1,33 +1,0 @@
-package com.example.scenchive.member.exception;
-
-import com.example.scenchive.global.exception.BaseExceptionType;
-import org.springframework.http.HttpStatus;
-
-public enum MemberExceptionType implements BaseExceptionType {
-    NOT_FOUND_MEMBER(602, HttpStatus.OK, "회원 정보가 없습니다.");
-
-    private int errorCode;
-    private HttpStatus httpStatus;
-    private String errorMessage;
-
-    MemberExceptionType(int errorCode, HttpStatus httpStatus, String errorMessage) {
-        this.errorCode = errorCode;
-        this.httpStatus = httpStatus;
-        this.errorMessage = errorMessage;
-    }
-
-    @Override
-    public int getErrorCode() {
-        return this.errorCode;
-    }
-
-    @Override
-    public HttpStatus getHttpStatus() {
-        return this.httpStatus;
-    }
-
-    @Override
-    public String getErrorMessage() {
-        return this.errorMessage;
-    }
-}
