--- Scenchive/src/main/java/com/example/scenchive/member/service/NotCorrespondingEmailException.java
@@ -1,10 +1,0 @@
-package com.example.scenchive.member.service;
-
-public class NotCorrespondingEmailException extends Exception {
-    public NotCorrespondingEmailException() {
-    }
-
-    public NotCorrespondingEmailException(String message) {
-        super(message);
-    }
-}
